fix(ops): drop restoration-mechanism touch entirely (maintainer-lens WOULD-CLOSE again — go fully minimal)
A third maintainer-lens pass (independent codex exec, same checklist) still
returned WOULD-CLOSE on the narrowed version: extending #2044's existing,
unmodified restoration mechanism to a second column was judged to still
count as touching write-back/critical-path semantics (checklist #2), even
though the mechanism itself was untouched. It also flagged the PR body's
evidence for the leak fix itself (not just the restoration mirror) as
insufficiently explicit (checklist #5).

This PR now touches exactly one thing: src/core/ops/pages.ts, stripping
private takes/facts fences from BOTH compiled_truth and timeline for
remote readers (unchanged from the first commit). No changes to
import-file.ts or facts-fence.ts at all — restoration semantics are
completely untouched by this PR.

Consequence, disclosed honestly in the PR body and issue #4546: stripping
timeline for the first time here widens (but does not newly invent) the
pre-existing #2044 round-trip hazard to a second column, without adding
restoration coverage for it yet. That coverage — together with the
row-level merge fix for #2044's two pre-existing gaps in compiled_truth
(P1 mixed-fence, P2 world-only-deletion) — is being filed as a separate,
independent follow-up PR against already-merged code, so this leak fix
can be evaluated as a single, minimal logical change.

Test changes: drop the entire "#2044 restoration mirrored to timeline"
describe block (3 tests) — restoration isn't touched by this PR anymore,
so there's nothing here to test. Layer A/B tests unchanged.

Verified: bun run verify (54/54) green; typecheck clean;
privacy-strip-and-forget.test.ts 17/17. Red-green on the actual leak
fix itself (not just the restoration mirror, per the evidence-bar
critique): reverting pages.ts to pristine upstream master reproduces the
exact leak (2 tests fail, private row appears in both the `timeline`
field and the serialized `content`/`text` round-trip field); restoring
the fix returns both to green.
