+++ src/core/facts-fence.ts
@@ -467,49 +467,3 @@ export function stripFactsFence(body: string, opts: StripFactsFenceOpts = {}): s
   const replacement = renderFactsTable(kept);
   return body.slice(0, beginIdx) + replacement + body.slice(endIdx + FACTS_FENCE_END.length);
 }
-
-/** Slice out a `FACTS_FENCE_BEGIN...FACTS_FENCE_END` block verbatim, or null if absent/unterminated. */
-export function extractFactsFenceBlock(body: string): string | null {
-  const beginIdx = body.indexOf(FACTS_FENCE_BEGIN);
-  if (beginIdx === -1) return null;
-  const endIdx = body.indexOf(FACTS_FENCE_END, beginIdx + FACTS_FENCE_BEGIN.length);
-  if (endIdx === -1) return null;
-  return body.slice(beginIdx, endIdx + FACTS_FENCE_END.length);
-}
-
-export function replaceOrAppendFactsFence(body: string, fenceBlock: string): string {
-  const beginIdx = body.indexOf(FACTS_FENCE_BEGIN);
-  if (beginIdx !== -1) {
-    const endIdx = body.indexOf(FACTS_FENCE_END, beginIdx + FACTS_FENCE_BEGIN.length);
-    if (endIdx !== -1) {
-      return body.slice(0, beginIdx) + fenceBlock + body.slice(endIdx + FACTS_FENCE_END.length);
-    }
-  }
-
-  const sep = body.endsWith('\n') ? '\n' : '\n\n';
-  return `${body}${sep}## Facts\n\n${fenceBlock}\n`;
-}
-
-/**
- * #2044: remote get_page intentionally strips private facts rows. A
- * documented get_page -> edit -> put_page round-trip can therefore arrive
- * with an empty/missing Facts fence even though the existing page still has
- * canonical fence rows. Preserve the old fence in that narrow case so the
- * system-of-record markdown is not truncated by the privacy boundary.
- * Shared by import-file.ts's compiled_truth and timeline call sites.
- */
-export function restoreZeroedFactsFence(incomingBody: string, existingBody: string): string {
-  const incomingFacts = parseFactsFence(incomingBody);
-  const existingFacts = parseFactsFence(existingBody);
-  const existingFenceBlock = extractFactsFenceBlock(existingBody);
-  if (
-    incomingFacts.facts.length === 0 &&
-    incomingFacts.warnings.length === 0 &&
-    existingFacts.warnings.length === 0 &&
-    existingFacts.facts.length > 0 &&
-    existingFenceBlock
-  ) {
-    return replaceOrAppendFactsFence(incomingBody, existingFenceBlock);
-  }
-  return incomingBody;
-}
